backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.25.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
limits==3.6.0
redis==5.0.1
shap==0.43.0
locust==2.17.0
python-dotenv==1.0.0
gunicorn==21.2.0

